Extension Syncing
We will be using a number of MMF2 extensions. To keep those synced up,
we will keep a folder which mimics the folder structure of the one in
Program Files (since Windows won't let Git automatically sync with
things in the Program Files folder). When we add or change an extension,
it will automatically sync with the local mirror. BUT you will have to
manually merge this folder with the one in Program Files.

To do this, simply open one window for your Program Files folder
(probably Program Files (x86)). Then open another and navigate to your
repository folder (mine is C:\gamedev\Bike Prototypes and yours should
be too). Make sure your MMF folders are named the same thing (if they're
not, rename the one in Program Files to be "Multimedia Fusion 2" without
the quotes so they match and therefore always merge properly). Drag the
Multimedia Fusion 2 folder from your repository folder to your Program
Files folder. Assuming your MMF folders are named the same thing, it
should ask you if you want to merge the folders. You do. You can also
click "do this for the rest of the items" or whatever.

Since you don't already have the extension I'm including here, that will
probably be the end of it except that Windows will need your permission
as Administrator to copy/create files and folders in the Program Files
folder. Give it permission for all the items.

The next time you go to merge folders though -- say, when we start using
another extension or an effect -- it will ask you if you want to Replace
or Don't Move the duplicate files. It doesn't really matter which you do
(certainly don't choose "Copy and Rename"), except in the case of us
updating an extension.